requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
Babel==2.9.1
certifi==2022.12.7
charset-normalizer==3.1.0
frozenlist==1.3.3
greenlet==2.0.2
idna==3.4
magic-filter==1.0.9
multidict==6.0.4
Pillow==9.4.0
psycopg2-binary==2.9.5
python-dateutil==2.8.2
python-decouple==3.8
pytz==2022.7.1
six==1.16.0
SQLAlchemy==2.0.6
swagger-client==1.0.0
typing_extensions==4.5.0
urllib3==1.26.15
yarl==1.8.2
